rpi/feature-extractor/requirements.txt
numpy==1.23.5
antropy==0.1.5
nats-py==2.3.1
pandas==2.1.1
PyWavelets==1.4.1
scipy==1.10.0
statsmodels==0.14.0